+++ core/src/main/java/arez/spy/TransactionStartedEvent.java
@@ -50,7 +50,7 @@ public final class TransactionStartedEvent
   public void toMap( @Nonnull final Map<String, Object> map )
   {
     map.put( "type", "TransactionStarted" );
-    map.put( "transaction", getName() );
+    map.put( "name", getName() );
     map.put( "mutation", isMutation() );
     final ObserverInfo tracker = getTracker();
     map.put( "tracker", null == tracker ? null : tracker.getName() );