+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * **\[core\]** In the spy events, convert the `duration` parameter into integers as no duration should
   exceed the size of an integer.
+* **\[core\]** In spy events, change the key used to identify the source of the event from source specific
+  string (i.e. `computed`, `observable`, `observer`, etc.) to use a generic key `name`.
 
 ### [v0.111](https://github.com/arez/arez/tree/v0.111) (2018-10-19)
 [Full Changelog](https://github.com/arez/arez/compare/v0.110...v0.111)